update: more api

+++ update.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 
+if [ ! -d ".git" ]; then
+  echo "Please deploy using Git."
+  exit 1
+fi
+
+if ! command -v git &> /dev/null; then
+    echo "Git is not installed! Please install git and try again."
+    exit 1
+fi
+
+git config --global --add safe.directory $(pwd)
 git fetch --all && git reset --hard origin/master && git pull origin master
 rm -rf composer.lock composer.phar
 wget https://github.com/composer/composer/releases/latest/download/composer.phar -O composer.phar
